# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Figma Challenge",
        "description": "Documentation des schémas des requêtes",
        "contact": {
            "name": "Louqmane Inoussa"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/challenges": {
            "get": {
                "tags": [
                    "Challenges"
                ],
                "summary": "Lister toutes les éditions",
                "operationId": "df61bd7503e62a2251f358931e0f3a5f",
                "responses": {
                    "200": {
                        "description": "Liste des éditions retournée avec succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Challenges"
                ],
                "summary": "Créer une nouvelle édition",
                "operationId": "63ffd18fb5f5371ec322250145cbb046",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "edition",
                                    "status"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Figma Challenge 2025"
                                    },
                                    "edition": {
                                        "type": "string",
                                        "example": "2025"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Une édition spéciale..."
                                    },
                                    "cover": {
                                        "type": "string",
                                        "example": "cover.jpg"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "draft",
                                            "open",
                                            "closed"
                                        ],
                                        "example": "open"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-08-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Challenge créé avec succès"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/challenges/{id}": {
            "get": {
                "tags": [
                    "Challenges"
                ],
                "summary": "Voir une édition précise",
                "operationId": "ed8928d8cd2720e0dcce6cd0e29767f1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'édition",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Challenge trouvé"
                    },
                    "404": {
                        "description": "Challenge introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Challenges"
                ],
                "summary": "Modifier une édition",
                "operationId": "8f55380ce146081226d3559b05c43570",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du challenge à modifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Nouvel intitulé"
                                    },
                                    "edition": {
                                        "type": "string",
                                        "example": "2025"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description mise à jour"
                                    },
                                    "cover": {
                                        "type": "string",
                                        "example": "new-cover.jpg"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "draft",
                                            "open",
                                            "closed"
                                        ],
                                        "example": "closed"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-09-30"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Challenge mis à jour"
                    },
                    "404": {
                        "description": "Introuvable"
                    },
                    "422": {
                        "description": "Validation échouée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Challenges"
                ],
                "summary": "Supprimer une édition",
                "operationId": "8621e1614a29d01b58dd2c4384d0afba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'édition à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Challenge supprimé"
                    },
                    "404": {
                        "description": "Introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/contact-messages": {
            "get": {
                "tags": [
                    "Contact Messages"
                ],
                "summary": "Lister tous les messages de contact",
                "operationId": "0be20a7e3c5a4c1eaa68c758a3e7eb1c",
                "responses": {
                    "200": {
                        "description": "Liste des messages"
                    }
                }
            },
            "post": {
                "tags": [
                    "Contact Messages"
                ],
                "summary": "Soumettre un message de contact",
                "operationId": "5e271348f1c0fa1a6a13264a1d2a8c9e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "message"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jane@example.com"
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "Bonjour, j'ai une question concernant..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Message enregistré"
                    }
                }
            }
        },
        "/api/contact-messages/{id}": {
            "get": {
                "tags": [
                    "Contact Messages"
                ],
                "summary": "Afficher un message de contact",
                "operationId": "d04356f09b379f43b7ba80c2858cf35f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message trouvé"
                    },
                    "404": {
                        "description": "Message non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Contact Messages"
                ],
                "summary": "Supprimer un message de contact",
                "operationId": "376bba76f4c0bc69f74b952c1cd9c449",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message supprimé"
                    }
                }
            }
        },
...
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}